NB-174 : refactor - S3 업로드된 삭제 기능 개선
1. deleteImages() 메소드 개선
 - 변경 전 : objectKey로 단일 삭제하는 deleteImage()를 반복해서 호출, 여러 번 통신함
 - 변경 후 : deleteObjects()를 호출하여 여러 objectKey를 한 번의 통신으로 삭제
 - 다수의 S3 객체를 단일 API 호출로 삭제하도록 변경

2. deleteImages() 메소드 호출 부분 개선
 - 파라미터로 전달될 List를 확인하여 빈 리스트인 경우 호출하지 않음

+++ src/main/java/com/soyeon/nubim/domain/album/AlbumService.java
@@ -95,7 +95,9 @@ public class AlbumService {
 		Map<Integer, String> newPhotoUrls = albumUpdateRequestDto.getPhotoUrls();
 		List<String> deletedS3ObjectKeys = getDeletedS3ObjectKeys(currentPhotoUrls, newPhotoUrls);
 
-		s3ImageDeleter.deleteImages(deletedS3ObjectKeys);
+		if (!deletedS3ObjectKeys.isEmpty()) {
+			s3ImageDeleter.deleteImages(deletedS3ObjectKeys);
+		}
 		album.setPhotoUrls(newPhotoUrls);
 
 		albumRepository.deleteLocationsByAlbumId(albumId);
